refactor(browser): Move email to shared
